+++ src/main/java/com/sweetrpg/catherder/common/block/entity/CatBowlBlockEntity.java
@@ -1,10 +1,10 @@
-package com.sweetrpg.catherder.common.block.tileentity;
+package com.sweetrpg.catherder.common.block.entity;
 
 import com.sweetrpg.catherder.api.feature.FoodHandler;
 import com.sweetrpg.catherder.common.entity.CatEntity;
 import com.sweetrpg.catherder.common.inventory.container.FoodBowlContainer;
 import com.sweetrpg.catherder.common.lib.Constants;
-import com.sweetrpg.catherder.common.registry.ModTileEntityTypes;
+import com.sweetrpg.catherder.common.registry.ModBlockEntityTypes;
 import com.sweetrpg.catherder.common.util.InventoryUtil;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -49,7 +49,7 @@ public class CatBowlBlockEntity extends PlacedBlockEntity implements MenuProvide
     public int timeoutCounter;
 
     public CatBowlBlockEntity(BlockPos pos, BlockState blockState) {
-        super(ModTileEntityTypes.CAT_BOWL.get(), pos, blockState);
+        super(ModBlockEntityTypes.CAT_BOWL.get(), pos, blockState);
     }
 
     public static void tick(Level level, BlockPos pos, BlockState blockState, BlockEntity blockEntity) {
@@ -57,8 +57,8 @@ public class CatBowlBlockEntity extends PlacedBlockEntity implements MenuProvide
             return;
         }
 
-        //Only run update code every 5 ticks (0.25s)
-        if(++bowl.timeoutCounter < 5) {return;}
+        // Only run update code every 5 ticks (0.25s)
+        if(++bowl.timeoutCounter < 5) { return; }
 
         List<CatEntity> catList = bowl.level.getEntitiesOfClass(CatEntity.class, new AABB(pos).inflate(5, 5, 5));