+++ src/main/java/com/sweetrpg/catherder/common/registry/ModBlockEntityTypes.java
@@ -1,11 +1,10 @@
 package com.sweetrpg.catherder.common.registry;
 
 import com.sweetrpg.catherder.api.CatHerderAPI;
-import com.sweetrpg.catherder.common.block.CatTreeBlock;
-import com.sweetrpg.catherder.common.block.tileentity.CatTreeBlockEntity;
-import com.sweetrpg.catherder.common.block.tileentity.CatBowlBlockEntity;
-import com.sweetrpg.catherder.common.block.tileentity.PetDoorBlockEntity;
-import com.sweetrpg.catherder.common.lib.Constants;
+import com.sweetrpg.catherder.common.block.entity.CatTreeBlockEntity;
+import com.sweetrpg.catherder.common.block.entity.CatBowlBlockEntity;
+import com.sweetrpg.catherder.common.block.entity.LitterboxBlockEntity;
+import com.sweetrpg.catherder.common.block.entity.PetDoorBlockEntity;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -15,7 +14,7 @@ import net.minecraftforge.registries.RegistryObject;
 
 import java.util.function.Supplier;
 
-public class ModTileEntityTypes {
+public class ModBlockEntityTypes {
 
     public static final DeferredRegister<BlockEntityType<?>> TILE_ENTITIES = DeferredRegister.create(ForgeRegistries.Keys.BLOCK_ENTITY_TYPES, CatHerderAPI.MOD_ID);
 
@@ -23,6 +22,7 @@ public class ModTileEntityTypes {
     public static final RegistryObject<BlockEntityType<PetDoorBlockEntity>> PET_DOOR = register("pet_door", PetDoorBlockEntity::new, ModBlocks.PET_DOOR);
 
     public static final RegistryObject<BlockEntityType<CatBowlBlockEntity>> CAT_BOWL = register("cat_bowl", CatBowlBlockEntity::new, ModBlocks.CAT_BOWL);
+    public static final RegistryObject<BlockEntityType<LitterboxBlockEntity>> LITTERBOX = register("litter_box", LitterboxBlockEntity::new, ModBlocks.LITTERBOX);
 
     private static <T extends BlockEntity> RegistryObject<BlockEntityType<T>> register(final String name, final BlockEntityType.BlockEntitySupplier<T> sup, Supplier<? extends Block> validBlock) {
         return register(name, () -> BlockEntityType.Builder.of(sup, validBlock.get()).build(null));