+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `[FIX]` Fixed missing texture issue when putting a collar on the cat.
 - `[FIX]` Add cat tree and cat toy to Item Physics.
 - `[NEW]` Holding the shift key while looking at the cat will now show maximum hunger as well as current.
+- `[NEW]` Added litter scoop and new configuration option to enable litter box maintenance.
 
 ## 0.0.75-alpha